+++ app/smartbrain_3000/web/_app/immutable/chunks/CzbnXWEv.js
@@ -1,0 +1,1 @@
+import{q as e}from"./Ch8MiCKH.js";import{n as t}from"./D7E9GVN2.js";var n=e({count:0});async function r(){try{n.count=(await t.unseenScheduleUpdates()).count}catch{}}export{n,r as t};