+++ app/smartbrain_3000/web/_app/immutable/entry/start.BwGR8WCK.js
@@ -1,0 +1,1 @@
+import{a as e,n as t}from"../chunks/BSFtaxTs.js";export{e as load_css,t as start};